feat: distinguish failed and ignored for JUnit (#719) (#879)
Previously all not passed scan results were treated as a failure by the
JUnit formatter. This was a problem for e.g. tfsec, since with a JUnit
result viewer there was no way to distinguish between ignored tests and
actually failed ones.

Now "ignored" scan results are marked as "skipped".

Additionally, a test for passed scans has been added as well. It already
worked as intended, but so far we had no test to ensure this.

Co-authored-by: Liam Galvin <liam.galvin@aquasec.com>